Forbes Marshall/requirements.txt
customtkinter==5.2.1
Pillow>=10.2.0
bcrypt>=4.1.2
pyserial>=3.5
